requirements.txt
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.5
EasyProcess==0.3
idna==3.2
lxml==4.6.3
numpy==1.21.2
Pillow==8.3.2
pynput==1.7.3
python-slugify==5.0.2
PyVirtualDisplay==0.2.4
requests==2.26.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
text-unidecode==1.3
urllib3==1.26.6
